code refactor/methods names adjusted/ directory structure changed
